[WIP] move styles into seperate file for consistency
