+++ apps/aws-amplify/src/App.tsx
@@ -1,9 +1,8 @@
 import { useMemo } from 'react';
 import { locations } from '@contentful/app-sdk';
+import { useSDK } from '@contentful/react-apps-toolkit';
 import ConfigScreen from './locations/ConfigScreen';
 import Sidebar from './locations/Sidebar';
-
-import { useSDK } from '@contentful/react-apps-toolkit';
 
 const ComponentLocationSettings = {
   [locations.LOCATION_APP_CONFIG]: ConfigScreen,
